+++ client/components/containers/menu/ReleaseMenu.vue
@@ -11,8 +11,8 @@
 import Vue, { PropType } from 'vue';
 
 import ContextMenu, { MenuItem } from '~/components/parts/menu/ContextMenu.vue';
-import ArtistMenu, { Props as ArtistMenuProps } from '~/components/parts/menu/ArtistMenu.vue';
-import PlaylistMenu, { Props as PlaylistMenuProps } from '~/components/containers/menu/PlaylistMenu.vue';
+import ArtistLinkMenu, { Props as ArtistLinkMenuProps } from '~/components/parts/menu/ArtistLinkMenu.vue';
+import AddItemToPlaylistMenu, { Props as AddItemToPlaylistMenuProps } from '~/components/containers/menu/AddItemToPlaylistMenu.vue';
 import ShareMenu, { Props as ShareMenuProps } from '~/components/parts/menu/ShareMenu.vue';
 import { MENU_BACKGROUND_COLOR, Z_INDEX_OF } from '~/variables';
 import { App } from '~~/types';
@@ -46,9 +46,9 @@ export default Vue.extend({
       const artistPageItem = () => {
         const { artistList } = this.release;
         if (artistList.length > 1) {
-          const props: ArtistMenuProps = { artistList };
+          const props: ArtistLinkMenuProps = { artistList };
           return {
-            component: ArtistMenu,
+            component: ArtistLinkMenu,
             props,
           };
         }
@@ -81,14 +81,14 @@ export default Vue.extend({
 
       const addItemToPlaylist = () => {
         const uriList = this.release.trackList.map((track) => track.uri);
-        const props: PlaylistMenuProps = {
+        const props: AddItemToPlaylistMenuProps = {
           name: this.release.name,
           uriList,
           artistList: this.release.artistList,
         };
 
         return {
-          component: PlaylistMenu,
+          component: AddItemToPlaylistMenu,
           props,
         };
       };
