+++ client/components/parts/menu/ArtistLinkMenu.vue
